* Changed bug report message for undefined script. * Added/Fixed some scripts for Ark introduction. * Added some scripts for FriendStory questline * Refactored some grand athenaeum scripts. * Refactored luminous introduction. * Added luminous questline until 2nd job. * Refactored drop data (still need do changes in rates so it will be updated constantly). * Added maps to unity portal. * Updated luminous new larkness system. * Fixed some NPE. * Updated party&friends operations. * Added skill prepare handling for each job. * Fixed 1st job advance for luminous dark&light.

+++ src/main/java/net/swordie/ms/client/character/commands/AdminCommands.java
@@ -826,7 +826,7 @@ public class AdminCommands {
                 level++;
                 chr.setStat(Stat.level, level);
                 Map<Stat, Object> stats = new HashMap<>();
-                stats.put(Stat.level, (byte) level);
+                stats.put(Stat.level,  (int) level);
                 stats.put(Stat.exp, (long) 0);
                 chr.getClient().write(WvsContext.statChanged(stats));
                 chr.getJobHandler().handleLevelUp();
